Refactor verbose stats to reduce memory consumption
Fixes #11

This is unfortunately a breaking change from 4379f504
in that any in-progress encodes made using 4379f504
will not resume with this commit.
Any in-progress encodes from before 4379f504
will resume successfully (but verbose stats will not be accurate).
If you are currently on 4379f504, finish in-progress encodes
before upgrading.
Given the small amount of time 4379f504 was on master for,
and the small user base of this application,
I'm considering this to be an acceptable breakage.