Switch the progress file to use msgpack
Since the progress format is already breaking
due to other changes,
this seems like an ideal time to swap before a release.
There's no need for the progress file to be human readable,
so might as well use the smallest, fastest format available
for these monstrous progress files that can be multiple MB
for full encodes.