Support distributed encoding
Adds rav1e-worker, which acts as a server on a remote encoding machine
and listens to connections from clients.
The client, part of rav1e-by-gop, negotiates with the server,
the server notifies the client as it has slots available,
and the client sends one segment at a time to the server.
The server can have multiple worker threads,
and a client can use as many threads as the server will allow.
A server can also serve multiple clients at once,
which it does via a FIFO queue.
The queue only holds slot requests,
so the client does not send frames until the server signals a slot is ready.
The client will send to the first worker that signals a slot is available.
This keeps memory usage down on both the client and server,
and streamlines the threading process so the end of an encode is not stalled
waiting for a specific remote worker,
if there are other workers or local threads available.

Communication is done using websockets, and is insecure by default,
but can be configured to use TLS.
Unfortunately this does not currently work with self-signed certs.
Client-server authentication is done using a simple password system;
a password is configured on the server,
and then the client must be configured to use the same password for this server.

The frame sending process is quite upload-heavy.
The input frames are sent zstd-compressed to try and alleviate this,
however results will still be best
if workers are on a LAN or other connection with a fast upload.
However, I found that even with a 10Mbps upload from my home connection,
the remote workers still were efficient enough to be worth using.