Use two-pass per segment for bitrate limit, for better quality
Actually, I hate rav1e's two-pass API.
It's bad. Like, really hard to use.
The only way I got it working was by copy-pasting
what was in rav1e's CLI. And now I think I understand
how it works, but it's still a lot of code that could
be significantly less code. I really shouldn't need
to know about the internal structure of the pass data
in order to use the API. We should really just implement
serde for it and have a simple API that is impossible
to use wrong. Because the current one is very, very
easy to use wrong.