Write the whole segment to file at once
Store the whole encoded segment in memory
to reduce the number of kernel calls needed
to write to the file.
I suspect a significant portion of the memory usage occurring
is due to the high volume of kernel calls due to file writes,
as smem reports an alarmingly high kernel memory usage
when rav1e-by-gop is running.